Restructured project
NodeJS allows easy version control and package handling as well as allows for a structured framework for seamless full stack development. Express allows for easy translation from JS to http, the article explains it more. Express-handlebars is also a really nice tool with express when it comes to handling HTML for the front end, also better explained in the article: https://blog.risingstack.com/node-hero-tutorial-getting-started-with-node-js/